\name{hybrid-package}
\alias{hybrid-package}
\alias{hybrid}
\docType{package}
\title{
\packageTitle{hybrid}
}
\description{
\packageDescription{hybrid}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hybrid}
\packageIndices{hybrid}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hybrid}

Maintainer: \packageMaintainer{hybrid}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
